Init v0.1 of module
Module fucntional
