+++ .gitignore
@@ -1,6 +1,4 @@
 code.zip
-Notes
-sudo.template
 # Local .terraform directories
 **/.terraform/*
 
@@ -15,7 +13,7 @@ crash.log
 # .tfvars files are managed as part of configuration and so should be included in
 # version control.
 #
-# example.tfvars
+*.tfvars
 
 # Ignore override files as they are usually used to override resources locally and so
 # are not checked in
