--- Commands
@@ -1,5 +1,0 @@
-terraform init
-terraform plan
-terraform apply -var-file=variables.tfvars -auto-approve
-terraform destroy
-terraform refresh
