--- README.md
@@ -1,2 +1,0 @@
-# sudo-terraform-module-template
-Template repository for Terraform modules
