+++ code/crhelper/__pycache__/resource_helper.cpython-37.pyc
