+++ outputs.tf
@@ -1,9 +1,4 @@
-output "lambda_arn" {
-  value       = module.cf_custom_lambda_sns.lambda_arn
-  description = "Use this ARN for Lambda backed CloudFormation Custom Resource."
-}
-
-output "sns_topic_arn" {
-  value       = module.cf_custom_lambda_sns.sns_topic_arn
+output "cf_backed_sns_arn" {
+  value       = aws_sns_topic.cf_backed_sns.arn
   description = "Use this ARN for SNS backed CloudFormation Custom Resource."
 }
