+++ variables.tf
@@ -1,2 +1,5 @@
-variable "shared_credentials_file" {}
-variable "profile" {}
+variable "email_address" {
+    type = string
+    description = "Email address for subscribing to the SNS topic."
+    default = "saif.ali@sudoconsultants.com"
+}
