--- variables.tfvars
@@ -1,2 +1,0 @@
-shared_credentials_file = "/c/Users/ssaz9/.aws/credentials"
-profile = "terraform"
